Change property value from "1" to true in phpcs.xml.dist
The ignoreUnusedValuesWhenOnlyKeysAreUsedInForeach is a boolean value.
It should be set to true. If it is set to the string "1" Slevomat will
break with the following message:

PHP Fatal error:  Uncaught TypeError: Return value of
[...]\UnusedVariableSniff::isValueInForeachAndErrorIsIgnored() must be
of the type bool, string returned in
[...]/vendorr/slevomat/coding-standard/SlevomatCodingStandard/Sniffs/Variables/UnusedVariableSniff.php:528

This effects Slevomat 7.0.19.

Change-Id: I8a0277bb3ba977c76677f70aa57c079cd3becfab
Issue-Link: https://project.pronovix.net/issues/4735